Update 15-workflow-sets.Rmd to use `option_add(param_info = ...)`
This is in reference to [tidymodels/workflows#49](https://github.com/tidymodels/workflowsets/issues/49#issuecomment-847127601) where the parameter object must be specified as `param_info` when passing it via `option_add` to a `workflow_set`. `param` is a partial that will likely work for passing the argument to `param_info` appropriately, but I incorrectly assumed that it might also be named `parameters`. This might prevent that confusion for others.